any player can kill any enemy whether master or slave (initial, read comments)
Godot is screwy in that you cant RPC pass everything in a node.... passing 'self' through rpc call does pass everything. player 'stats' exists on enemy side, but is not populated through the RPC call. basically empty variable.
changed RPC call from self to stats.dmg(), this passed through ok.
EXP passing was commented out as was not necesary for testing and was in the way.