server & client picked up items appear in inventory (weapons only, see comments)
as godot cant pass a node reference i collected all item data in a dictionary, passed to client, then had client "item" node generate the item with the passed data itself. seems to work.

Wepons only, see if you like the solution or prefer something else. if yes ill extend it to armour and wearables. if no, modify it or roll it back.