Item syncing bug (see comments, important)
example.

Client was master of lvl2 and lvl3.
Client was currently on lvl3. (therefor grid set to level 3)
Server was on lvl2.
server killed something.
item was generating on lvl3 as clients "grid" was set to lvl3.

Fix was to get master of grid level (client) to generate item on grid level of the slave (server) was on.

pressing 'e' had the same effect, enemies generated on lvl 3 where grid master currently was. (not fixed yet)

i think this may also cause problem with grid_sync. it will sync whatever everyones active grid is, instead of the specific level grid.